flatpak: Re-order modules and fix pandoc build
For flatpaks the last module should usually be the application being
installed. In this case, uberwriter. In addition, the build process
for pandoc made an assumption that /app/bin would be present after
building the uberwriter module, but since that wasn't the case /app/bin/pandoc
must be created in advance.

Fixes issues where the flatpak build wouldn't run due to "missing" python modules.