Add script used to generate theme colors
Theme colors are originally based on GitHub's theme.

This script (slightly) offsets them to match the color pallete of the
specified theme automatically, and prints them out.

While the adjustments are meant to be minor, color contrast should still
be verified.

Finally, a few themes should not be automated in this way. For example,
the high contrast themes are meant to be purely black & white.